+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## - 2017-06-21
+- Display options section added
+- Vertical panning script added in display options
+
 ## - 2017-06-16
 - Existing bars transition scripts renamed as bars transition #1.
 - Bars transition #2 implemented.